Add support for conventional commit
This commit adds support for conventional commit to the tool. A new flag
`--conventional-commit` has been added to the CLI, which when set to
true, generates commit messages that are compliant with the conventional
commit standard. The `MessageConfig` struct has been updated to include
a new field `ConventionalCommitCompliant` to enable this feature. The
`GetCommitMessage` function has been updated to include the conventional
commit content in the message if the flag is set to true. Additionally,
a new function `getExpectedMessage` has been added to return the
expected commit message based on the style and conventional commit flag.